+++ goty-client/src/components/results/Summary/Summary.module.scss
