+++ goty-client/src/components/results/Summary/Summary.tsx
@@ -7,6 +7,12 @@ import { GameOfTheYearResult } from '../../../models/gameOfTheYearResult'
 import { GameResult } from '../../../models/gameResult'
 import { Results } from '../../../models/results'
 import styles from './Summary.module.scss'
+import {
+  selectIsLoading,
+  selectResults,
+} from '../../../state/results/selectors'
+import { Loading } from '../../Loading'
+import { Card } from '../../controls/Card/Card'
 
 const stringHeaders: Header<string>[] = [
   {
@@ -56,12 +62,23 @@ const gotyHeaders: Header<GameOfTheYearResult>[] = [
   },
 ]
 
-export interface SummaryProps {
-  results: Results
-}
+const isEmptyResults = (results: Results) =>
+  Object.values(results).every((value) => value.length === 0)
 
-export const Summary = ({ results }: SummaryProps) => {
+export const Summary = () => {
+  const isLoading = useSelector(selectIsLoading)
+  const results = useSelector(selectResults)
   const { year } = useSelector(selectProperties)
+  if (isLoading) {
+    return <Loading />
+  }
+  if (isEmptyResults(results)) {
+    return (
+      <Card>
+        <h2>No results yet</h2>
+      </Card>
+    )
+  }
   return (
     <>
       <Respondents rows={results.participants} />
